requirements.txt
h5py==3.9.0
matplotlib==3.5.1
monai==0.8.1
nibabel==3.2.2
numpy==1.21.5
pydicom==2.4.3
pytorch_lightning==1.4.9
PyYAML==6.0
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
torchinfo==1.8.0
torchvision==0.10.1+cu111
tqdm==4.62.3
umap_learn==0.5.2
wandb==0.12.11
lpips==0.1.4
protobuf==3.20.*
plotly==5.6.0
opencv-python==4.8.0.76
